+++ src/main/java/edu/ie3/simona/api/ontology/em/RequestEmCompletion.java
@@ -6,7 +6,7 @@
 
 package edu.ie3.simona.api.ontology.em;
 
-import java.util.Optional;
+import java.util.OptionalLong;
 
 /**
  * Request send to SIMONA to finish the em service for the given tick.
@@ -14,9 +14,9 @@ import java.util.Optional;
  * @param tick for which the em service should be finished
  * @param maybeNextTick option for the next tick
  */
-public record RequestEmCompletion(long tick, Optional<Long> maybeNextTick)
+public record RequestEmCompletion(long tick, OptionalLong maybeNextTick)
     implements EmDataMessageFromExt {
   public RequestEmCompletion(long tick) {
-    this(tick, Optional.empty());
+    this(tick, OptionalLong.empty());
   }
 }