+++ src/main/java/edu/ie3/simona/api/ontology/simulation/CompletionMessage.java
@@ -6,12 +6,12 @@
 
 package edu.ie3.simona.api.ontology.simulation;
 
-import java.util.Optional;
+import java.util.OptionalLong;
 
 /**
  * Message that is returned to SIMONA
  *
  * @param nextActivation The tick that the external simulation would like to be activated at again
  */
-public record CompletionMessage(Optional<Long> nextActivation)
+public record CompletionMessage(OptionalLong nextActivation)
     implements ControlResponseMessageFromExt {}